design: 로그인 페이지 css 변경

+++ backend/nest/src/socket/socket.service.ts
@@ -1,11 +1,12 @@
-import { HttpException, HttpStatus, Inject, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import * as fs from 'fs';
 import { DeviceService } from './../device/device.service';
 import { SocketLoginDto } from './socket.dto';
 import { DeviceCreateDto } from 'src/device/device-req.dto';
 import { SentenceService } from 'src/sentence/sentence.service';
 import { TtsService } from 'src/tts/tts.service';
-import { RedisService } from 'src/redis/redis.service';
+import { FileService } from './../file/file.service';
+import { SentenceCreateDto } from 'src/sentence/sentence-req.dto';
 
 @Injectable()
 export class SocketService {
@@ -13,7 +14,7 @@ export class SocketService {
     private readonly deviceService: DeviceService,
     private readonly sentenceService: SentenceService,
     private readonly ttsService: TtsService,
-    private readonly redisService: RedisService,
+    private readonly fileService: FileService,
   ){}
 
   // device 정보 + socket id 저장
@@ -24,8 +25,8 @@ export class SocketService {
     // 파라미터 없음
     if (serial_number==null) throw new HttpException("plz serial_number", HttpStatus.BAD_REQUEST);
 
-    // redis에 소켓 id 저장
-    await this.redisService.set(serial_number, clientId)
+    // 소켓 id 저장 해야함
+
 
     // 처음온 연결인 경우
     if (device == null){
@@ -45,43 +46,53 @@ export class SocketService {
     return result;
   }
 
-  /** stt-> tts : stt 받아서 tts로 응답 */
-  async stt(text: string, talk_id: string, base64Data: string): Promise<string>{
-    const saveFilePath = "./upload/2024-02-01/" + "푸른이와의 대화.mp3"
+  /** stt-> tts : stt 받아서 tts로 return */
+  async stt(text: string, talk_id: number, base64Data: string): Promise<string>{
+    const today = this.fileService.getToday()
+    const filePath = "./upload/talk/" + today + "/"
+    if (!fs.existsSync(filePath)) fs.mkdir(filePath, (e)=>{if (e) throw e})
+
+    let nextSentenceId = await this.sentenceService.nestSentenceId(talk_id)
+    const saveFilePath =  filePath + talk_id + "-" + nextSentenceId + ".mp3"
     const decodedBuffer = Buffer.from(base64Data, 'base64');
     fs.writeFileSync(saveFilePath, decodedBuffer);
 
+    const sentenceDto = new SentenceCreateDto()
+    sentenceDto.content = text
+    sentenceDto.audio = saveFilePath
+    sentenceDto.sentence_DTN = today as unknown as Date
+    sentenceDto.talker = "user"
+    sentenceDto.talk_id = talk_id
+    await this.sentenceService.save(sentenceDto)
     // gpt api
     const answerText = await this.sentenceService.answer(text)
     
-    const filePath = "./upload/2024-01-30/" + "ETA.wav"
+    const uploadFilePath = filePath + talk_id + "-" + (nextSentenceId+1) + ".wav"
     // message -> tts
-    await this.ttsService.tts(answerText, filePath)
-
+    await this.ttsService.tts(answerText, uploadFilePath)
+    
     // client.emit
     const content = await new Promise<Buffer>((resolve, reject) => {
-      fs.readFile(filePath, (err, data) => {
+      fs.readFile(uploadFilePath, (err, data) => {
         if (err) {
-        reject(err)
+          reject(err)
         } else {
-        resolve(data)
+          resolve(data)
         }
       })
     })
-
+    
     // text, answerText 파일 저장 -> redis
     console.log(text)
+    
+    const sentenceDto2 = new SentenceCreateDto()
+    sentenceDto2.content = answerText
+    sentenceDto2.audio = uploadFilePath
+    sentenceDto.sentence_DTN = today as unknown as Date
+    sentenceDto2.talker = "ai"
+    sentenceDto2.talk_id = talk_id
+    await this.sentenceService.save(sentenceDto2)
 
-    let sentenceId = this.redisService.incr(`${talk_id}:sentenceId`)
-    // '{talkId}:{sentenceId}':key, text:value
-    await this.redisService.set(`${talk_id}:${sentenceId}`, text)
-    await this.redisService.set(`${talk_id}:${sentenceId}:url`, saveFilePath)
-    // sentenceId ++
-    sentenceId = this.redisService.incr(`${talk_id}:sentenceId`)
-    await this.redisService.set(`${talk_id}:${sentenceId}`, answerText)
-    await this.redisService.set(`${talk_id}:${sentenceId}:url`, filePath)
-    // if sentenceId//2 == 0 => child 
-    // else AI
     return Buffer.from(content).toString('base64')
   }
 }
